Use dynamodb namespace
