+++ lib/event_tracer/dynamo_db/client.rb
@@ -1,0 +1,9 @@
+module EventTracer
+  module DynamoDB
+    class Client
+      def self.call
+        Aws::DynamoDB::Client.new
+      end
+    end
+  end
+end
