+++ lib/event_tracer/dynamo_db/default_processor.rb
@@ -1,0 +1,15 @@
+module EventTracer
+  module DynamoDB
+    class DefaultProcessor
+      def call(log_type, action:, message:, args:)
+        args.merge(
+          timestamp: Time.now.utc.iso8601(6),
+          action: action,
+          message: message,
+          log_type: log_type,
+          app: EventTracer::Config.config.app_name
+        )
+      end
+    end
+  end
+end
