--- lib/event_tracer/dynamo_db_logging/dynamo_db_client.rb
@@ -1,7 +1,0 @@
-module EventTracer
-  class DynamoDBClient
-    def self.call
-      Aws::DynamoDB::Client.new
-    end
-  end
-end
